requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
django-filter==2.4.0
django-telegram-login==0.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
docker==5.0.0
idna==3.2
kombu==5.1.0
npm==0.1.1
optional-django==0.1.0
Pillow==8.3.1
prompt-toolkit==3.0.19
PyJWT==2.1.0
pytz==2021.1
pywin32==227
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.2.1
